requirements.txt
# H200 GPU Cluster Workshop - Python Dependencies
# Install with: pip install -r requirements.txt

# PyTorch is installed separately with CUDA support:
# pip install torch torchvision --index-url https://download.pytorch.org/whl/cu121

# Additional dependencies
matplotlib>=3.5.0
pillow>=9.0.0
numpy>=1.21.0
